changes for auth code runs locally but from harvest app doesn't waiting for response to continue

+++ prodeagle/harvest.go
@@ -5,7 +5,7 @@ import (
 	"appengine/memcache"
 	"bytes"
 	"encoding/json"
-	"net/http"
+	//"net/http"
 	"strconv"
 	"strings"
 	"time"
@@ -28,35 +28,68 @@ import (
   "time": 1319752200,           // the current time on the server modulo 60
   "version": 1.0                // the version of the protocol. so far always 1.0
 }
+
+
+
+{
+	"all_data_inaccurate":false,
+	"ms_of_data_lost":1622,
+	"time":1371494040,
+	"version":"1.0",
+	"counters":{
+		"xydddssdfgd":{
+
+		},
+		"xydddssfhfghfgh":{
+
+		}
+	}
+}
+
+
+{
+	"all_data_inaccurate":false,
+	"counters":{
+		"xydddssdfgd":{
+			"1371494460":1
+		},
+		"xydddssfhfghfgh":{
+			"1371494460":2
+		}
+	},
+	"ms_of_data_lost":1060,
+	"time":1371494460,
+	"version":1}
+
 **/
 
 type counterHarvest struct {
 	All_data_inaccurate bool                        `json:"all_data_inaccurate"`
+	Counters            map[string]map[string]int64 `json:"counters"`
 	Ms_of_data_lost     int64                       `json:"ms_of_data_lost"`
 	Time                int64                       `json:"time"`
-	Version             string                      `json:"version"`
-	Counters            map[string]map[string]int64 `json:"counters"`
+	Version             float32                     `json:"version"`
 }
 
 const (
-	max_look_back_time  int64  = 3600 //1h
-	max_clock_skew      int64  = 60
-	min_slot_size       int64  = 60
-	sep                 string = "_"
-	version             string = "1.0"
-	max_memcache_server int    = 1024
-	lost_data_check     string = "lost_data_check"
+	max_look_back_time  int64   = 3600 //1h
+	max_clock_skew      int64   = 60
+	min_slot_size       int64   = 60
+	sep                 string  = "_"
+	version             float32 = 1.0
+	max_memcache_server int     = 1024
+	lost_data_check     string  = "lost_data_check"
 )
 
 //harvest stored counters
 //if last_time is not set try to harvest counters from last hour
-func Harvest(w http.ResponseWriter, r *http.Request) {
+func Harvest(c appengine.Context, sLastHarvestTime string, prodCall bool) []byte {
 	startTime := time.Now()
-	dc := appengine.NewContext(r)
-	c, _ := appengine.Namespace(dc, namespace)
+	//dc := appengine.NewContext(r)
+	//c, _ := appengine.Namespace(dc, namespace)
 	//check if there could be lost counters since last harvest
 	all_data_inaccurate := wasDataLost(c, true)
-	sLastHarvestTime := r.FormValue("last_time")
+	//sLastHarvestTime := r.FormValue("last_time")
 	lastHarvestTime := time.Now().Unix()
 	if sLastHarvestTime != "" {
 		lastHarvestTime, _ = strconv.ParseInt(sLastHarvestTime, 10, 64)
@@ -69,7 +102,10 @@ func Harvest(w http.ResponseWriter, r *http.Request) {
 		counters[name] = make(map[string]int64)
 	}
 	cmNames := createMemCacheNames(counterNames)
-	keys := make([]string, len(counters_name)*2, len(counters_name)*3)
+	var keys []string
+	if prodCall {
+		keys = make([]string, len(counters_name)*2, len(counters_name)*3)
+	}
 	c.Infof("lastharvesttime is: %v", lastHarvestTime)
 	c.Infof("currenttime is: %v", currentTime)
 	c.Infof("slot is: %v", slot)
@@ -79,7 +115,9 @@ func Harvest(w http.ResponseWriter, r *http.Request) {
 		items, _ := memcache.GetMulti(c, cmNames(sslot))
 		for key, item := range items {
 			buf := bytes.NewBuffer(item.Value)
-			keys = append(keys, key)
+			if prodCall {
+				keys = append(keys, key)
+			}
 			cn := strings.Split(key, sep)
 			counters[cn[1]][cn[0]], _ = strconv.ParseInt(buf.String(), 10, 64)
 		}
@@ -92,12 +130,16 @@ func Harvest(w http.ResponseWriter, r *http.Request) {
 	b, err := json.Marshal(harvest)
 	if err != nil {
 		c.Errorf("Harvest - json.Marshal(%#v) %s ", counters, err)
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
+		//http.Error(w, err.Error(), http.StatusInternalServerError)
+		return nil
 	}
-	//delete harvested counters
-	memcache.DeleteMulti(c, keys)
-	w.Write(b)
+	if prodCall {
+		//delete harvested counters
+		c.Infof("deleting all counters")
+		memcache.DeleteMulti(c, keys)
+	}
+	//w.Write(b)
+	return b
 }
 
 func wasDataLost(c appengine.Context, reset bool) bool {
